Separate components from containers
