+++ src/actions/actions_users.js
@@ -91,3 +91,7 @@ export function getUserProfile(wallet, isCurrentUser) {
               })
   }
 }
+
+export function updateUserProfile() {
+
+}
